Codechange: Uniformly name Tile parameters t
When a function accepts a `Tile` as its fist argument, uniformly name
this parameter `t`.

Examining the existing codebase, `Tile t` seems to be the preferred
naming scheme:

```
$ find ./src/ -regex '.*\.\(cpp\|hpp\|h\)' | xargs grep -E '\(Tile t' | grep -E --only-matching 'Tile t[^,)]*' | sort | uniq --count
    290 Tile t
      1 Tile test_tile
     61 Tile tile
```

This is a codehealth no-op change to improve uniformity of the codebase.
One copy&paste error in the documentation (likely due to the
inconsistency, in TryReserveTrack, the @pre IsPlainRailTile) is also fixed.